Move .of_enumerable, .of_enumerator tests to separate file.
